+++ src/bar.rs
@@ -1,5 +1,6 @@
 //! The bar module handles the i3bar protocol.
 
+use crate::common::Color;
 use serde::Serialize;
 
 /// Represents the output of a block.
@@ -8,13 +9,16 @@ use serde::Serialize;
 #[derive(Debug, Serialize)]
 pub struct BlockOutput {
     /// The text to display in the block. Example: "BAT 42% 02:12".
-    full_text: String,
+    pub full_text: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub background: Option<Color>,
 }
 
 impl BlockOutput {
     pub fn new(full_text: &str) -> Self {
         Self {
             full_text: full_text.to_string(),
+            background: None,
         }
     }
 }