+++ src/blocks/battery.rs
@@ -2,6 +2,7 @@
 
 use crate::bar::BlockOutput;
 use crate::blocks::Block;
+use crate::common::Color;
 use crate::config::BatteryConfig;
 use crate::os::battery::*;
 
@@ -10,6 +11,10 @@ pub struct BatteryBlock {
     format: String,
     /// The 0-based index of the battery to monitor.
     index: Option<u8>,
+    /// The background color when the level is critical.
+    bg_critical: Color,
+    /// The percentage from which the level is considered critical.
+    critical_level: u8,
 }
 
 impl BatteryBlock {
@@ -17,15 +22,18 @@ impl BatteryBlock {
         BatteryBlock {
             format: config.format.to_string(),
             index: config.index,
+            bg_critical: config.bg_critical,
+            critical_level: config.critical_level,
         }
     }
 }
 
 impl Block for BatteryBlock {
     fn get_output(&self) -> BlockOutput {
-        let rem_percentage: String = match get_battery_level(self.index) {
-            Ok(l) => format!("{}", l),
-            Err(e) => e.to_string(),
+        
+        let (rem_percentage_s, rem_percentage): (String, u8) = match get_battery_level(self.index) {
+            Ok(l) => (format!("{}", l), l),
+            Err(e) => (e.to_string(), 0),
         };
         let bat_state: String = match get_battery_state() {
             Ok(s) => format!("{}", s),
@@ -37,9 +45,14 @@ impl Block for BatteryBlock {
         };
         let out = self
             .format
-            .replace("{rem_percent}", &rem_percentage)
+            .replace("{rem_percent}", &rem_percentage_s)
             .replace("{chr_state}", &bat_state)
             .replace("{rem_time}", &rem_time);
-        BlockOutput::new(&format!("{}", out))
+
+        let mut block = BlockOutput::new(&format!("{}", out));
+        if rem_percentage <= self.critical_level {
+            block.background = Some(self.bg_critical);
+        }
+        block
     }
 }